- move makeRandomMove to shared.go file - make move for services counselor
