+++ pkg/testdatagen/scenario/devseed.go
@@ -16,10 +16,6 @@ import (
 	"net/http/httptest"
 	"time"
 
-	fakedata "github.com/transcom/mymove/pkg/fakedata_approved"
-
-	"github.com/transcom/mymove/pkg/random"
-
 	"github.com/stretchr/testify/mock"
 
 	paymentrequestop "github.com/transcom/mymove/pkg/gen/primeapi/primeoperations/payment_request"
@@ -3438,107 +3434,6 @@ func createHHGNeedsServicesCounseling(db *pop.Connection) {
 			Status:                models.MTOShipmentStatusSubmitted,
 			RequestedPickupDate:   &requestedPickupDate,
 			RequestedDeliveryDate: &requestedDeliveryDate,
-		},
-	})
-}
-
-func createRandomMove(db *pop.Connection, possibleStatuses []models.MoveStatus, allDutyStations []models.DutyStation, dutyStationsInGBLOC []models.DutyStation, assertions testdatagen.Assertions) {
-	randDays, err := random.GetRandomInt(366)
-	if err != nil {
-		log.Panic(fmt.Errorf("Unable to generate random integer for submitted move date"), zap.Error(err))
-	}
-	submittedAt := time.Now().AddDate(0, 0, randDays*-1)
-
-	if assertions.ServiceMember.Affiliation == nil {
-		randomAffiliation, err := random.GetRandomInt(5)
-		if err != nil {
-			log.Panic(fmt.Errorf("Unable to generate random integer for affiliation"), zap.Error(err))
-		}
-		assertions.ServiceMember.Affiliation = &[]models.ServiceMemberAffiliation{
-			models.AffiliationARMY,
-			models.AffiliationAIRFORCE,
-			models.AffiliationNAVY,
-			models.AffiliationCOASTGUARD,
-			models.AffiliationMARINES}[randomAffiliation]
-	}
-
-	dutyStationCount := len(allDutyStations)
-	if assertions.Order.OriginDutyStationID == nil {
-		// We can pick any origin duty station not only one in the office user's GBLOC
-		if *assertions.ServiceMember.Affiliation == models.AffiliationMARINES {
-			randDutyStaionIndex, err := random.GetRandomInt(dutyStationCount)
-			if err != nil {
-				log.Panic(fmt.Errorf("Unable to generate random integer for duty station"), zap.Error(err))
-			}
-			assertions.Order.OriginDutyStation = &allDutyStations[randDutyStaionIndex]
-			assertions.Order.OriginDutyStationID = &assertions.Order.OriginDutyStation.ID
-		} else {
-			randDutyStaionIndex, err := random.GetRandomInt(len(dutyStationsInGBLOC))
-			if err != nil {
-				log.Panic(fmt.Errorf("Unable to generate random integer for duty station"), zap.Error(err))
-			}
-			assertions.Order.OriginDutyStation = &dutyStationsInGBLOC[randDutyStaionIndex]
-			assertions.Order.OriginDutyStationID = &assertions.Order.OriginDutyStation.ID
-		}
-	}
-
-	if assertions.Order.NewDutyStationID == uuid.Nil {
-		randDutyStaionIndex, err := random.GetRandomInt(dutyStationCount)
-		if err != nil {
-			log.Panic(fmt.Errorf("Unable to generate random integer for duty station"), zap.Error(err))
-		}
-		assertions.Order.NewDutyStation = allDutyStations[randDutyStaionIndex]
-		assertions.Order.NewDutyStationID = assertions.Order.NewDutyStation.ID
-	}
-
-	randomFirst, randomLast := fakedata.RandomName()
-	assertions.ServiceMember.FirstName = &randomFirst
-	assertions.ServiceMember.LastName = &randomLast
-
-	orders := testdatagen.MakeOrderWithoutDefaults(db, assertions)
-
-	if assertions.Move.SubmittedAt == nil {
-		assertions.Move.SubmittedAt = &submittedAt
-	}
-
-	if assertions.Move.Status == "" {
-		randStatusIndex, err := random.GetRandomInt(len(possibleStatuses))
-		if err != nil {
-			log.Panic(fmt.Errorf("Unable to generate random integer for move status"), zap.Error(err))
-		}
-		assertions.Move.Status = possibleStatuses[randStatusIndex]
-
-		if assertions.Move.Status == models.MoveStatusServiceCounselingCompleted {
-			counseledAt := submittedAt.Add(3 * 24 * time.Hour)
-			assertions.Move.ServiceCounselingCompletedAt = &counseledAt
-		}
-	}
-	move := testdatagen.MakeMove(db, testdatagen.Assertions{
-		Move:  assertions.Move,
-		Order: orders,
-	})
-
-	laterRequestedPickupDate := submittedAt.Add(60 * 24 * time.Hour)
-	laterRequestedDeliveryDate := laterRequestedPickupDate.Add(7 * 24 * time.Hour)
-	testdatagen.MakeMTOShipment(db, testdatagen.Assertions{
-		Move: move,
-		MTOShipment: models.MTOShipment{
-			ShipmentType:          models.MTOShipmentTypeHHG,
-			Status:                models.MTOShipmentStatusSubmitted,
-			RequestedPickupDate:   &laterRequestedPickupDate,
-			RequestedDeliveryDate: &laterRequestedDeliveryDate,
-		},
-	})
-
-	earlierRequestedPickupDate := submittedAt.Add(30 * 24 * time.Hour)
-	earlierRequestedDeliveryDate := earlierRequestedPickupDate.Add(7 * 24 * time.Hour)
-	testdatagen.MakeMTOShipment(db, testdatagen.Assertions{
-		Move: move,
-		MTOShipment: models.MTOShipment{
-			ShipmentType:          models.MTOShipmentTypeHHG,
-			Status:                models.MTOShipmentStatusSubmitted,
-			RequestedPickupDate:   &earlierRequestedPickupDate,
-			RequestedDeliveryDate: &earlierRequestedDeliveryDate,
 		},
 	})
 }
